Add files via upload
This project incl. a content based recom system which gives movie suggestion based on the user entered movie name. It uses Tfidfvectorization to transform text into numbers and uses cosine similarity to create a similarity score. Based in the index of the movie given by user , it compares its similarityscore to the others and produces similar movies